fix and add solution according comments
