+++ src/Root.tsx
@@ -7,27 +7,25 @@ import { PeoplePage } from './pages/PeoplePage';
 
 import { HashRouter } from 'react-router-dom';
 import { App } from './App';
-import { MyProvider } from './components/dataContext/dataContext';
+// import { MyProvider } from './components/dataContext/dataContext';
+import { NotFoundPage } from './pages/NotFoundPage';
 
 export const Root = () => (
   <HashRouter>
-    <MyProvider>
-      <Routes>
-        <Route path="/" element={<App />}>
-          <Route index element={<HomePage />} />
-          <Route path="home" element={<Navigate to="/" replace />} />
+    {/* <MyProvider> */}
+    <Routes>
+      <Route path="/" element={<App />}>
+        <Route index element={<HomePage />} />
+        <Route path="home" element={<Navigate to="/" replace />} />
 
-          <Route path="people">
-            <Route index element={<PeoplePage />} />
-            <Route path=":slug" element={<PeoplePage />} />
-            <Route
-              path="*"
-              element={<h1 className="title">Page not found</h1>}
-            />
-          </Route>
-          <Route path="*" element={<h1 className="title">Page not found</h1>} />
+        <Route path="people">
+          <Route index element={<PeoplePage />} />
+          <Route path=":slug" element={<PeoplePage />} />
+          <Route path="*" element={<NotFoundPage />} />
         </Route>
-      </Routes>
-    </MyProvider>
+        <Route path="*" element={<NotFoundPage />} />
+      </Route>
+    </Routes>
+    {/* </MyProvider> */}
   </HashRouter>
 );
