+++ src/pages/PeoplePage.tsx
@@ -1,11 +1,35 @@
 import 'bulma/css/bulma.css';
-import { useContext } from 'react';
+import { useEffect, useState } from 'react';
 import { Table } from '../components/Table/Table';
-import { dataContext } from '../components/dataContext/dataContext';
+// import { dataContext } from '../components/dataContext/dataContext';
 import { Loader } from '../components/Loader';
+import { preparePeopleData } from '../utils/preparePeopleData';
+import { Person } from '../types';
+import { getPeople } from '../api';
 
 export const PeoplePage: React.FC = () => {
-  const { dataFromServer, errorMessage, isLoading } = useContext(dataContext)!;
+  // const { dataFromServer, errorMessage, isLoading } = useContext(dataContext)!;
+  const [dataFromServer, setDataFromServer] = useState<Person[]>();
+  const [isLoading, setIsLoading] = useState<boolean>(false);
+  const [errorMessage, setErrorMessage] = useState<string | null>(null);
+
+  useEffect(() => {
+    setDataFromServer([]);
+    setErrorMessage(null);
+    setIsLoading(true);
+    getPeople()
+      .then((data: Person[]) => {
+        const peopleData = preparePeopleData(data);
+
+        setDataFromServer(peopleData);
+      })
+      .catch(() => {
+        setErrorMessage('Something went wrong');
+      })
+      .finally(() => {
+        setIsLoading(false);
+      });
+  }, []);
 
   if (isLoading) {
     return <Loader />;
@@ -23,5 +47,10 @@ export const PeoplePage: React.FC = () => {
     return <p data-cy="noPeopleMessage">There are no people on the server</p>;
   }
 
-  return <Table />;
+  return (
+    <>
+      <h1 className="title">People Page</h1>
+      <Table people={dataFromServer} isLoading={isLoading} />
+    </>
+  );
 };
